+++ src/refactoring/hooks/admin/index.ts
@@ -1,2 +1,3 @@
 export * from "./useProductForm";
 export * from "./useProductUpdateForm";
+export * from "./useCouponForm";
